feat: add private route /profile
